requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.2.2=pypi_0
attrs=22.2.0=pypi_0
blas=1.0=mkl
blinker=1.5=pypi_0
bottleneck=1.3.5=py310h9128911_0
branca=0.6.0=pypi_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
cachetools=5.3.0=pypi_0
certifi=2022.12.7=py310haa95532_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
click-plugins=1.1.1=pypi_0
cligj=0.7.2=pypi_0
colorama=0.4.6=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pyhd3eb1b0_0
decorator=5.1.1=pypi_0
dtaidistance=2.3.10=pypi_0
entrypoints=0.4=pypi_0
filelock=3.9.0=py310haa95532_0
fiona=1.9.1=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
folium=0.14.0=pypi_0
fonttools=4.39.0=pypi_0
freetype=2.12.1=ha860e81_0
geopandas=0.12.2=pypi_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
idna=3.4=pypi_0
importlib-metadata=6.0.0=pypi_0
intel-openmp=2021.4.0=haa95532_3556
jinja2=3.1.2=py310haa95532_0
jpeg=9e=h2bbff1b_1
jsonschema=4.17.3=pypi_0
kiwisolver=1.4.4=py310hd77b12b_0
lerc=3.0=hd77b12b_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib-base=3.7.1=py310h4ed8f06_0
mdurl=0.1.2=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py310h2bbff1b_0
mkl_fft=1.3.1=py310ha0764ea_0
mkl_random=1.2.2=py310h4ed8f06_0
mpmath=1.2.1=py310haa95532_0
munch=2.5.0=pypi_0
munkres=1.1.4=py_0
networkx=2.8.4=py310haa95532_0
numexpr=2.8.4=py310hd213c9f_0
numpy=1.24.2=pypi_0
numpy-base=1.23.5=py310h04254f7_0
openssl=1.1.1t=h2bbff1b_0
packaging=23.0=pypi_0
pandas=1.5.3=py310h4ed8f06_0
patsy=0.5.3=pypi_0
pillow=9.4.0=pypi_0
pip=23.0.1=py310haa95532_0
plotly=5.13.1=pypi_0
plotly-express=0.4.1=pypi_0
protobuf=3.20.3=pypi_0
pyarrow=11.0.0=pypi_0
pydeck=0.8.0=pypi_0
pygments=2.14.0=pypi_0
pympler=1.0.1=pypi_0
pyparsing=3.0.9=py310haa95532_0
pyproj=3.4.1=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.10.9=h966fe2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-decouple=3.8=pypi_0
pytorch=2.0.0=py3.10_cpu_0
pytorch-mutex=1.0=cpu
pytz=2022.7.1=pypi_0
pytz-deprecation-shim=0.1.0.post0=pypi_0
requests=2.28.2=pypi_0
rich=13.3.2=pypi_0
scipy=1.10.1=pypi_0
seaborn=0.12.2=py310haa95532_0
semver=2.13.0=pypi_0
setuptools=65.6.3=py310haa95532_0
shapely=2.0.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
sqlite=3.41.1=h2bbff1b_0
statsmodels=0.13.5=pypi_0
streamlit=1.20.0=pypi_0
streamlit-aggrid=0.3.4.post3=pypi_0
streamlit-folium=0.11.1=pypi_0
sympy=1.11.1=py310haa95532_0
tenacity=8.2.2=pypi_0
tk=8.6.12=h2bbff1b_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
typing-extensions=4.5.0=pypi_0
typing_extensions=4.4.0=py310haa95532_0
tzdata=2022.7=pypi_0
tzlocal=4.2=pypi_0
urllib3=1.26.15=pypi_0
validators=0.20.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
watchdog=2.3.1=pypi_0
wheel=0.38.4=py310haa95532_0
wincertstore=0.2=py310haa95532_2
xz=5.2.10=h8cc25b3_1
zipp=3.15.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
